Add warning/confirmation for reposting link topics
If a user tries to post a link topic that's been posted before, this
will add a warning below the form with info about the previous posts and
ask them to confirm that they want to re-post. Currently, this doesn't
have any sort of time/group/etc. restrictions and will just find
previous topics from all time.

The method for doing this is a big ugly to be able to handle both the
no-JS and with-JS posting methods, but it's not too bad.