Improve wrapping of comment post/edit times
Previously the posted/edited times of comments wrapped very poorly on
small screens. This improves it somewhat, making it so they won't wrap
"internally" but the edited time will go underneath the posted time.

More improvements are still needed, especially with the Link/Parent nav
links, but this is definitely better.