Show errors from Intercooler <button> requests
Previously, error messages were only shown inside <form> elements. If a
button hit an error (such as a 403 when trying to vote on a comment that
was deleted after page load), the button simply wouldn't work with no
indication of why.

This adds the error message into the <menu> containing the buttons, and
involved some reworking of the relevant JS and CSS.