Pin pyramid-session-redis version to 1.5.0
Version 1.5.1 of pyramid-session-redis has a change to how session IDs
are generated that will cause all existing sessions to be invalidated.
Before upgrading it would be best to set up some migration to be able to
keep the existing sessions, so that everyone doesn't need to log back
in.

There's some info in pyramid-session-redis's CHANGES file:
https://github.com/jvanasco/pyramid_session_redis/blob/master/CHANGES.md