Refactor tag list columns to use TypeDecorator
I've always been pretty unhappy with the ugly way tags were implemented,
using @hybrid_property and needing to do strange things all over the
place to deal with converting underscores to spaces and vice versa, as
well as other idiosyncracies.

There are still a few oddities here and there, but overall this is much
better.