Allow users to search their own posts
This actually involves the necessary backend changes for full comment
search, but I'm a little nervous about the privacy implications of that,
and don't want to enable it for everyone out of nowhere. So for now,
this just allows users to search *their own* comments or topics.

These views and templates are starting to get real ugly, and probably
need a major re-thinking soon.