Start using pip-tools and split dev dependencies
I'm going to start using pip-tools to manage dependencies:
https://github.com/jazzband/pip-tools

This makes updating the dependencies and virtualenv easier in a few
ways, and makes it simple to keep dev dependencies split out (so I can
stop installing them in production).

Now, to do a check and update all packages to their newest versions, the
main command is:

pip-compile --no-header --upgrade requirements.in

and again with requirements-dev.in to update that one as well. This will
update all the package versions in requirements.txt and
requirements-dev.txt. The virtualenv can then be updated to match those
versions by running:

pip-sync requirements.txt

(or requirements-dev.txt for dev environment). This currently needs to
be run with sudo, but I'm going to try to fix that shortly.