Vagrantfile: pin Salt version to 2019.2.3
The new version of Salt ("3000") seems to have a number of bugs,
including not being able to handle "unless" checks, which the Tildes
states use frequently. Because of this, creating a new dev environment
currently doesn't work. This pins Salt to the previous stable version
for now.

Here's the relevant bug for "unless" specifically:
https://github.com/saltstack/salt/issues/56131

And the overall release notes:
https://docs.saltstack.com/en/latest/topics/releases/3000.html