Don't post full "backlog" of scheduled topics
This will probably only ever be relevant in development environments,
but we don't want the topic scheduler to always post a full backlog of
scheduled topics when it hasn't run for a while. For example, if a dev
environment has a daily scheduled topic set up, but the VM is not
launched for a week, the next time the "post scheduled topics" cronjob
runs, it will post all 7 of the backlogged topics.

This commit changes the script so that it advances the schedule to the
next *future* occurrence, instead of continuing the backlog.